+++ example/src/scenes/Lamp/index.js
@@ -27,6 +27,13 @@ import ConicalGradientScene from './ConicalGradient';
 import ColorSelectorAnimationScene from './ColorSelectorAnimation';
 import MusicDrawerScene from './MusicDrawer';
 import CountdownSubPage from './CountdownSubPage';
+import TabsBarScene from './TabsBar';
+import AnimatedModalScene from './AnimatedModal';
+import NormalTimingScene from './NormalTiming';
+import SceneFunScene from './SceneFun';
+import AddTimerScene from './NormalTiming/addTimer';
+import RepeatScene from './NormalTiming/weeks';
+import TestWhiteScene from './NormalTiming/testWhite';
 
 export default class LampScene extends Component {
   get data() {
@@ -58,6 +65,14 @@ export default class LampScene extends Component {
   static ConicalGradient = ConicalGradientScene;
   static ColorSelectorAnimation = ColorSelectorAnimationScene;
   static CountdownSubPage = CountdownSubPage;
+  static TabsBar = TabsBarScene;
+  static AnimatedModal = AnimatedModalScene;
+  static NormalTiming = NormalTimingScene;
+  static SceneFun = SceneFunScene;
+  static AddTimer = AddTimerScene;
+  static Repeat = RepeatScene;
+  static TestWhite = TestWhiteScene;
+
   render() {
     return <TYFlatList contentContainerStyle={{ paddingTop: 16 }} data={this.data} />;
   }