+++ packages/tuya-panel-lamp-sdk/src/components/timer/normal-timing/__tests__/customTop.test.js
@@ -1,0 +1,16 @@
+import React from 'react';
+import { shallow } from 'enzyme';
+import CustomTopBar from '../components/CustomTopBar';
+
+describe('CustomTopBar components', () => {
+  it('basic render', () => {
+    const props = {
+      themeColor: '#fff',
+      title: '',
+      onSave() {},
+      onBack() {},
+    };
+    const wrapper = shallow(<CustomTopBar {...props} />);
+    expect(wrapper).toMatchSnapshot();
+  });
+});