+++ packages/tuya-panel-lamp-sdk/src/hooks/index.ts
@@ -1,0 +1,2 @@
+export { default as useEffectOnce } from './useEffectOnce';
+export { default as useSceneData } from './useScene';